requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9.9
pgvector>=0.1.8
langchain>=0.1.0
langchain-core>=0.1.0
langchain-text-splitters>=0.3.0
langchain-openai>=0.1.0
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
PyPDF2>=3.0.0
openai>=1.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
cryptography>=41.0.0